# OrderService/src/main/java/ru/dmitryobuhkoff/orderservice/model/dto/request/OrderWithoutProductDtoRequest.java
package ru.dmitryobuhkoff.orderservice.model.dto.request;

import com.fasterxml.jackson.annotation.JsonFormat;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.Date;

@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
public class OrderWithoutProductDtoRequest {

    @JsonProperty("article")
    private Integer article;

    @JsonProperty("dateSince")
    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
    private Date dateSince;

    @JsonProperty("dateBefore")
    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
    private Date dateBefore;
}